v1.474 mini correcion

+++ MyrranCore/src/DTOs/DTOsNet.java
@@ -166,19 +166,19 @@ public class DTOsNet
 
     public static class EliminarAura
     {
-        public byte auraID;
+        public int auraID;
         public EliminarAura() {}
         public EliminarAura(AuraI aura)
-        {   this.auraID = (byte)aura.getID(); }
+        {   this.auraID = aura.getID(); }
     }
 
     public static class ModificarAuraStacks
     {
-        public byte auraID;
+        public int auraID;
         public byte numStacks;
         public ModificarAuraStacks() {}
         public ModificarAuraStacks(AuraI aura)
-        {   this.auraID = (byte)aura.getID(); this.numStacks = (byte)aura.getStacks(); }
+        {   this.auraID = aura.getID(); this.numStacks = (byte)aura.getStacks(); }
     }
 
     // PC DTOS:
